% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{conservation_status}
\alias{conservation_status}
\title{conservation_status}
\usage{
conservation_status(species)
}
\arguments{
\item{species}{the species to provide}
}
\value{

}
\description{
conservation_status
}
\examples{
conservation_status("Anemone occidentalis - Carex nigricans")

}
