% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_map}
\alias{species_map}
\title{species_map}
\usage{
species_map(species)
}
\arguments{
\item{species}{the species of interest}
}
\value{
a map indicating all ecosections where the species is extant
}
\description{
species_map
}
\examples{

}
